generalize callback in entity
