+++ homeassistant/components/dio_chacon/cover.py
@@ -50,20 +50,6 @@ class DioChaconCover(DioChaconEntity, CoverEntity):
         | CoverEntityFeature.SET_POSITION
     )
 
-    async def async_added_to_hass(self) -> None:
-        """Register the callback for server side events."""
-        await super().async_added_to_hass()
-        self.dio_chacon_client.set_callback_device_state_by_device(
-            self.target_id, self.callback_device_state
-        )
-
-    def callback_device_state(self, data: dict[str, Any]) -> None:
-        """Receive callback for device state notification pushed from the server."""
-
-        _LOGGER.debug("Data received from server %s", data)
-        self._update_attr(data)
-        self.async_write_ha_state()
-
     def _update_attr(self, data: dict[str, Any]) -> None:
         """Recomputes the attributes values either at init or when the device state changes."""
         self._attr_available = data["connected"]
