+++ homeassistant/components/dio_chacon/entity.py
@@ -1,5 +1,6 @@
 """Base entity for the Dio Chacon entity."""
 
+import logging
 from typing import Any
 
 from dio_chacon_wifi_api import DIOChaconAPIClient
@@ -8,6 +9,8 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import Entity
 
 from .const import DOMAIN, MANUFACTURER
+
+_LOGGER = logging.getLogger(__name__)
 
 
 class DioChaconEntity(Entity):
@@ -37,3 +40,17 @@ class DioChaconEntity(Entity):
     def _update_attr(self, data: dict[str, Any]) -> None:
         """Recomputes the attributes values."""
         # method to be overridden by the child class.
+
+    async def async_added_to_hass(self) -> None:
+        """Register the callback for server side events."""
+        await super().async_added_to_hass()
+        self.dio_chacon_client.set_callback_device_state_by_device(
+            self.target_id, self.callback_device_state
+        )
+
+    def callback_device_state(self, data: dict[str, Any]) -> None:
+        """Receive callback for device state notification pushed from the server."""
+
+        _LOGGER.debug("Data received from server %s", data)
+        self._update_attr(data)
+        self.async_write_ha_state()
